fix alignment and hide for panels with custom toolbars
